spi.py: gigantic refactor
refactor Token to IToken so that Position (line and col) can be
contained in a Token

IToken is now an abstract base class that is implemented by several
*Tok subtypes. IToken also contains Position. IToken now has
abstract property value since all tokens are expected to have a value

Lexer has been refactored to generate fresh IToken subtypes
without looking into a reserved keyword token cache since position
is required and unique.  Also, get rid of extra EOF token at the end

IAST requires __str__ to be implemented in the subclass

Parser refactored to take into account all of the above

SemanticAnalyzer refactored to take into account all of the above

DecoSrcBuilder refactored to take into account all of the above

Interpreter refactored to take into account all of the above

added dynamic casting (as opposed to the typing.cast which is a static
cast). This custom cast() which does dynamic casting can take in an
iterable where the target value must be an instance of at least one
of the types in the iterable